chore(tasks): refactor task list views
