+++ lib/services/data/tasks/filtered/task_filter_state_provider.dart
@@ -118,78 +118,3 @@ final taskFilterStateProvider =
       ref.read(navigationServiceProvider).navigatorKey.currentContext!;
   return TaskFilterStateNotifier(context);
 });
-
-final filteredTasksProvider = StateProvider.autoDispose
-    .family<List<TaskModel>, String?>((ref, projectId) {
-  final tasks = ref
-          .watch(projectId == null
-              ? curUserViewableTasksStreamProvider
-              : tasksByProjectStreamProvider(projectId))
-          .value ??
-      [];
-
-  final filterState = ref.watch(taskFilterStateProvider);
-
-  final filteredTasks =
-      tasks.where((task) => filterState.filterCondition(task)).toList();
-
-  if (filterState.sortComparator != null) {
-    filteredTasks.sort(filterState.sortComparator!);
-  }
-
-  final curInlineCreatingTaskId = ref.watch(curInlineCreatingTaskIdProvider);
-
-  if (curInlineCreatingTaskId != null &&
-      !filteredTasks.any((t) => t.id == curInlineCreatingTaskId)) {
-    filteredTasks.add(tasks.firstWhere((t) => t.id == curInlineCreatingTaskId));
-  }
-
-  return filteredTasks;
-});
-
-final filteredTasksAndParentsProvider = StateProvider.autoDispose
-    .family<List<TaskModel>, String?>((ref, projectId) {
-  final tasks = ref
-          .watch(projectId == null
-              ? curUserViewableTasksStreamProvider
-              : tasksByProjectStreamProvider(projectId))
-          .value ??
-      [];
-
-  // Create a set of all task IDs we need to include (filtered tasks + their ancestors)
-  final Set<String> tasksToInclude = {};
-
-  final curInlineCreatingTaskId = ref.watch(curInlineCreatingTaskIdProvider);
-
-  if (curInlineCreatingTaskId != null) {
-    tasksToInclude.add(curInlineCreatingTaskId);
-  }
-
-  final filterState = ref.watch(taskFilterStateProvider);
-
-  // Add filtered tasks and their ancestors
-  for (final task in tasks) {
-    if (filterState.filterCondition(task)) {
-      tasksToInclude.add(task.id);
-
-      // Add all ancestors
-      var currentTask = task;
-      while (currentTask.parentTaskId != null) {
-        final parent =
-            tasks.firstWhere((t) => t.id == currentTask.parentTaskId);
-        tasksToInclude.add(parent.id);
-        currentTask = parent;
-      }
-    }
-  }
-
-  // Create a filtered list of tasks that pass the filter
-  final filteredTasks =
-      tasks.where((task) => tasksToInclude.contains(task.id)).toList();
-
-  if (filterState.sortComparator != null) {
-    filteredTasks.sort(filterState.sortComparator!);
-  }
-
-  return filteredTasks;
-});
