+++ lib/widgets/home/home_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:seren_ai_flutter/services/data/notes/widgets/notes_list_page.dart';
-import 'package:seren_ai_flutter/services/data/projects/widgets/project_overview/task_search_modal.dart';
+import 'package:seren_ai_flutter/services/data/tasks/filtered/task_search_modal.dart';
 import 'package:seren_ai_flutter/services/data/shifts/widgets/shifts_page.dart';
 import 'package:seren_ai_flutter/widgets/home/screens/cur_user_tasks_screen.dart';
 import 'package:seren_ai_flutter/widgets/home/screens/recent_updated_items_screen.dart';
